# Improve performance to call Mysql2::Result#each and Mysql2::Result#fields